\name{get_args}
\alias{get_args}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get all arguments of a call
}
\description{
This function can only be called within another function, it then returns a named list of all arguments in the super-function.
}
\usage{
get_args()
}
%- maybe also 'usage' for other objects documented here.
\value{
A named list of all arguments in the super-function.
}
\author{
Wannes Rosiers
}